Move files to appropriate folders
