+++ src/pages/MyProfile.js
@@ -1,8 +1,15 @@
+import { useSelector } from 'react-redux';
+import JoinedMissionsList from '../components/JoinedMissionsList';
+import styles from '../styles/MyProfile.module.css';
+import { selectsAllMissions } from '../redux/mission/misionSlice';
+
 function MyProfile() {
+  const { mission } = useSelector(selectsAllMissions);
+  const joinMission = mission.filter((mission) => mission.reservation === true);
+
   return (
-    <section>
-      <h2>My Profile</h2>
-      <div>MyProfile page content</div>
+    <section className={styles.profiler}>
+      <JoinedMissionsList data={joinMission} />
     </section>
   );
 }
