+++ src/styles/JoinedMissionsList.module.css
@@ -1,0 +1,7 @@
+.profilertable {
+  color: #000; /* placehoder styles for linter */
+}
+
+.tabledata {
+  color: #000; /* placehoder styles for linter */
+}
